server/webserver/requirements.txt
fastapi
uvicorn[standard]
asyncpg
paho-mqtt
python-dotenv
